[CHORE] Removed MoM assets and refactored Decorations component.
