+++ .env.sample
@@ -11,7 +11,6 @@ VITE_INVENTORY_CONTRACT=0xA0643dDDbaDf11ca201F6fBba55852960A50A8b2
 VITE_TOKEN_CONTRACT=0x2c597768BDff675660Ef1C4a16DfD2a34425DF22
 VITE_WISHING_WELL_CONTRACT=0x2a0fA94A095bAE0722B37f43B8A6becD85B9cCf2
 VITE_PAIR_CONTRACT=0xaE42EA9864C36820627068a11959915f817ce809
-VITE_MOM_CONTRACT=0x2cFdb89954991b69D6D4B2ec8b70f3b995F4E8Ec
 VITE_DONATION_ADDRESS=0x6D18a54E0fd87FCb84a0510A3eCd8855b7226715
 VITE_DISCORD_REDIRECT=http://localhost:3000
 VITE_RECAPTCHA_SITEKEY=6Lfqm6MeAAAAAFS5a0vwAfTGUwnlNoHziyIlOl1s
