+++ .github/workflows/testnet.yml
@@ -38,7 +38,6 @@ jobs:
           VITE_TOKEN_CONTRACT: "0x2c597768BDff675660Ef1C4a16DfD2a34425DF22"
           VITE_WISHING_WELL_CONTRACT: "0x2a0fA94A095bAE0722B37f43B8A6becD85B9cCf2"
           VITE_PAIR_CONTRACT: "0xaE42EA9864C36820627068a11959915f817ce809"
-          VITE_MOM_CONTRACT: "0x2cFdb89954991b69D6D4B2ec8b70f3b995F4E8Ec"
           VITE_TRADER_CONTRACT: "0x66e8429dE30D973d4E1E090620ccD5BeC8B31806"
           VITE_TREASURY_ADDRESS: "0x1DB57407EE80709D4d862fe81399FBB35B8B9586"
           VITE_BUMPKIN_MINTER_CONTRACT: "0xd775dB2F16915BdCF8f92A42786282F5F78Ad46E"
