+++ src/features/game/components/decorations/index.ts
@@ -1,0 +1,7 @@
+export * from "./Scarecrows";
+export * from "./Beavers";
+export * from "./Moles";
+export * from "./NyonStatue";
+export * from "./Trivia";
+export * from "./RockGolem";
+export * from "./WickerMan";
